refactor: clarify store namespaces
Move run and step store implementations under domain namespaces, leave root store modules as compatibility facades, and keep tests mirrored with the renamed production modules.
