+++ lib/squid_mesh/runs/store/persistence.ex
@@ -1,14 +1,14 @@
-defmodule SquidMesh.RunStore.Persistence do
+defmodule SquidMesh.Runs.Store.Persistence do
   @moduledoc """
   Write-side persistence helpers for workflow runs.
 
   These helpers keep record construction and serialization close to the
-  database-facing code while `SquidMesh.RunStore` continues to expose the
+  database-facing code while `SquidMesh.Runs.Store` continues to expose the
   public lifecycle API.
   """
 
   alias SquidMesh.Run
-  alias SquidMesh.RunStore.Serialization
+  alias SquidMesh.Runs.Store.Serialization
 
   # Replays intentionally drop step-derived context. Only reserved run-level
   # facts that describe how the run was started are copied into the new run.
